app/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.2.1"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.11"
dependencies = [
    "py-cord==2.6.1",
    "python-dotenv==1.2.1",
    # ---- Production requirement (always kept) ----
    "pydiscogs @ git+https://github.com/bpafoshizle/pydiscogs@main",
]

# ---- Development override -------------------------------------------------
# When you are actively editing the library, uncomment the line below.
# It tells uv to pull **pydiscogs** from the local checkout instead of the
# remote Git repo, and the `editable = true` flag keeps it in “‑e” mode.
[tool.uv.sources]
pydiscogs = { path = "../libs/pydiscogs", editable = true }

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88



[dependency-groups]
dev = [
    "black>=24.0.0",
    "flake8>=7.0.0",
    "isort>=5.0.0",
]
